[arm64-warnigs-IOS-207] Fixed warnings caused by implicit type conversio...

+++ src/PHAPIRequest.m
@@ -467,7 +467,7 @@ static NSString *const kPHHTTPHeaderContentType = @"Content-Type";
     if (kPHRequestStatusInProgress == self.requestStatus || kPHRequestStatusSucceeded ==
                 self.requestStatus)
     {
-        PH_DEBUG(@"Trying to re-send request with status = %u", self.requestStatus);
+        PH_DEBUG(@"Trying to re-send request with status = %lu", (unsigned long)self.requestStatus);
         return;
     }
